+++ CHANGELOG.md
@@ -51,3 +51,8 @@
 * Added the ability to scroll the bottomsheet with just one finger
 * Added oneFingerScrolling parameter
 
+## 0.1.7
+
+* Dart 3 & Flutter 3.10 compatibility
+* Added 'animationDuration' parameter to DraggableBottomSheet to change the animation duration of bottomsheet
+